New Legend Position Options, New Error Bar Options, & Glitch Fixes
Fixed issues which arose when changing error bars from 95% confidence intervals to standard errors, issues occurring on some versions of R when attempting to rename the levels of one's panel variable, and issues which arose and when creating violin plots after the user removed a given level using the subset() command in base R. The present version also includes new options for changing the position of the legend (showlegendabove=TRUE, showlegendbelow=TRUE, and showlegendleft=TRUE), as well as an option to change the error bars to reflect one standard deviation above/below the mean (seterrorbars="sd"). Additionally, the command setgrouplevels=c() has been added as an alternative to setlegendlevels=c() for setting the levels of a grouping (color) variable.