Showgridlines=FALSE Fixes and Text Auto-Rotation Fixes
Fixed an issue which arose when setting showgridlines to FALSE that caused a legend to appear when it wasn't necessary. Previous versions also included a feature which rotated the x-axis text automatically when the names of the levels of the x-axis variable used too many characters (to avoid overlapping text on the x-axis). This, however, sometimes caused issues. For now, this feature has been removed. Vertical labels can be set manually using the command showrotatedxlabels=TRUE.