Small Improvements/Bug Fixes
The present version includes a new command, showlegend=FALSE, allowing users to hide the color legend. Additionally, the commands setxaxislabel= and setyaxislabel= have been added as alternatives to setyaxistitle= and setxaxistitle= for specifying axis labels. Finally, for violin plots, the present version also includes some small bug fixes and the new commands setboxplotoutlinecolor= (alternatively setwhiskercolor=) for specifying the color of boxplot outlines and whiskers, as well as the command showcoloredboxplotoutlines=TRUE (alternatively, showcoloredwhiskers=TRUE) for generating outlines and whiskers which match the colors of violins.