Bug Fixes, Graph_violin Updates, and New Warning Messages
Fixed a bug in the graph_scatterplot() function; previously, the setsplitgroup command worked, but the splitgroup command did not. Also removed the requirement that 4 or more subjects be present in all cells of your data when using the graph_violin() function and setting showviolin to false (e.g., when just showing boxplots and/or data without violins). Finally, added new warning messages when the default jitter is too high given the range of your data.