Add files via upload
Added animations to the graph_bar() function. Also adjusted the default animation settings for graph_violin and graph_scatterplot, fixed various bugs when creating animations, and added additional customization options (allowing users to modify the transition times and time spent on each graph, adjust the font size of the animation heading, and so on).